addons: move collectd to an addon
Note that this removes the logstash collectd integration. This needed to
happen because the Logstash role has become a generic log sink role, and
collectd logs are taking up a huge amount of room in that store. We'll
need to find a storage solution for collectd at a later date